+++ Managementagent/models/WorkRequest.cs
@@ -89,7 +89,7 @@ namespace Oci.ManagementagentService.Models
         
         /// <value>
         /// The date and time the request was created, as described in
-        /// [RFC 3339](https://tools.ietf.org/rfc/rfc3339), section 14.29.
+        /// [RFC 3339](https://tools.ietf.org/rfc/rfc3339), section 5.6.
         /// 
         /// </value>
         /// <remarks>
@@ -101,7 +101,7 @@ namespace Oci.ManagementagentService.Models
         
         /// <value>
         /// The date and time the request was started, as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339),
-        /// section 14.29.
+        /// section 5.6.
         /// 
         /// </value>
         [JsonProperty(PropertyName = "timeStarted")]