+++ Managementagent/models/WorkRequestResource.cs
@@ -82,7 +82,7 @@ namespace Oci.ManagementagentService.Models
         
         /// <value>
         /// The date and time the request was created, as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339),
-        /// section 14.29.
+        /// section 5.6.
         /// 
         /// </value>
         [JsonProperty(PropertyName = "timeAccepted")]
@@ -90,7 +90,7 @@ namespace Oci.ManagementagentService.Models
         
         /// <value>
         /// The date and time the request was started, as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339),
-        /// section 14.29.
+        /// section 5.6.
         /// 
         /// </value>
         [JsonProperty(PropertyName = "timeStarted")]
@@ -98,7 +98,7 @@ namespace Oci.ManagementagentService.Models
         
         /// <value>
         /// The date and time the request was finished, as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339),
-        /// section 14.29.
+        /// section 5.6.
         /// 
         /// </value>
         [JsonProperty(PropertyName = "timeFinished")]