+++ Managementagent/requests/ListAvailabilityHistoriesRequest.cs
@@ -38,7 +38,7 @@ namespace Oci.ManagementagentService.Requests
         /// <value>
         /// Filter to limit the availability history results to that of time after the input time including the boundary record.
         /// Defaulted to current date minus one year.
-        /// The date and time to be given as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339), section 14.29.
+        /// The date and time to be given as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339), section 5.6.
         /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "timeAvailabilityStatusEndedGreaterThan")]
@@ -47,7 +47,7 @@ namespace Oci.ManagementagentService.Requests
         /// <value>
         /// Filter to limit the availability history results to that of time before the input time including the boundary record
         /// Defaulted to current date.
-        /// The date and time to be given as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339), section 14.29.
+        /// The date and time to be given as described in [RFC 3339](https://tools.ietf.org/rfc/rfc3339), section 5.6.
         /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "timeAvailabilityStatusStartedLessThan")]