+++ Oce/models/CreateOceInstanceDetails.cs
@@ -165,6 +165,12 @@ namespace Oci.OceService.Models
         public System.Nullable<LicenseType> InstanceLicenseType { get; set; }
         
         /// <value>
+        /// disaster recovery paired ragion name
+        /// </value>
+        [JsonProperty(PropertyName = "drRegion")]
+        public string DrRegion { get; set; }
+        
+        /// <value>
         /// Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
         /// Example: {&quot;bar-key&quot;: &quot;value&quot;}
         /// </value>