+++ Oce/models/OceInstanceSummary.cs
@@ -214,6 +214,12 @@ namespace Oci.OceService.Models
         public System.Nullable<LifecycleDetails> LifecycleDetails { get; set; }
         
         /// <value>
+        /// disaster recovery paired ragion name
+        /// </value>
+        [JsonProperty(PropertyName = "drRegion")]
+        public string DrRegion { get; set; }
+        
+        /// <value>
         /// An message describing the current state in more detail. For example, can be used to provide actionable information for a resource in Failed state.
         /// </value>
         [JsonProperty(PropertyName = "stateMessage")]