+++ Oce/models/UpdateOceInstanceDetails.cs
@@ -65,6 +65,19 @@ namespace Oci.OceService.Models
         public System.Collections.Generic.List<string> AddOnFeatures { get; set; }
         
         /// <value>
+        /// Details of the current state of the instance lifecycle
+        /// </value>
+        [JsonProperty(PropertyName = "lifecycleDetails")]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public System.Nullable<LifecycleDetails> LifecycleDetails { get; set; }
+        
+        /// <value>
+        /// disaster recovery paired ragion name
+        /// </value>
+        [JsonProperty(PropertyName = "drRegion")]
+        public string DrRegion { get; set; }
+        
+        /// <value>
         /// Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
         /// Example: {&quot;bar-key&quot;: &quot;value&quot;}
         /// </value>