+++ Opsi/requests/SummarizeDatabaseInsightResourceUtilizationInsightRequest.cs
@@ -235,5 +235,24 @@ namespace Oci.OpsiService.Requests
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "vmclusterName", Oci.Common.Http.CollectionFormatType.Multi)]
         public System.Collections.Generic.List<string> VmclusterName { get; set; }
+        
+        /// <value>
+        /// Filter by one or more cdb name.
+        /// 
+        /// </value>
+        [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "cdbName", Oci.Common.Http.CollectionFormatType.Multi)]
+        public System.Collections.Generic.List<string> CdbName { get; set; }
+        
+        /// <value>
+        /// Percent value in which a resource metric is considered highly utilized.
+        /// </value>
+        [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "highUtilizationThreshold")]
+        public System.Nullable<int> HighUtilizationThreshold { get; set; }
+        
+        /// <value>
+        /// Percent value in which a resource metric is considered low utilized.
+        /// </value>
+        [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "lowUtilizationThreshold")]
+        public System.Nullable<int> LowUtilizationThreshold { get; set; }
     }
 }